fix: decouple getAuthErrorMessage from Firebase Auth SDK import chain
Login.tsx imported getAuthErrorMessage from firebase/authService, which
eagerly calls getAuth(app) at module load time. This pulled the real
Firebase Auth SDK initialization into MainApp.test.tsx's import graph
even though AuthContext is mocked there, causing the test suite to fail
in CI (no REACT_APP_FIREBASE_* env vars set) with:
  FirebaseError: Firebase: Error (auth/invalid-api-key)

Extract the pure error-message-mapping logic into
src/utils/authErrorMessage.ts, which has no Firebase dependency.
authService.ts re-exports it for backward compatibility; Login.tsx now
imports it directly from the new module.
